# Conflicts:
#	src/data/src/entity/groups.rs
#	src/data/src/entity/roles.rs

+++ src/data/src/migration/inserts.rs
@@ -39,7 +39,7 @@ use crate::entity::webauthn::PasskeyEntity;
 use crate::entity::webids::WebId;
 use crate::events::event::Event;
 use cryptr::EncValue;
-use hiqlite_macros::params;
+use hiqlite::macros::params;
 use rauthy_common::is_hiqlite;
 use rauthy_error::ErrorResponse;
 use std::cmp::max;
@@ -598,7 +598,7 @@ VALUES ($1, $2, $3)"#;
         for b in data_before {
             DB::pg_execute(
                 sql_2,
-                &[&b.client_id, &b.action.to_string(), &(b.retry_count as i32)],
+                &[&b.client_id, &b.action.to_string(), &b.retry_count],
             )
             .await?;
         }
