+++ src/schedulers/src/scim_tasks.rs
@@ -18,7 +18,7 @@ pub async fn scim_task_retry() {
 
     loop {
         // We want to randomize the sleep because this scheduler should run on all cluster members.
-        // This increases the chance opf success in case of a network segmentation.
+        // This increases the chance of success in case of a network segmentation.
         let millis = get_rand_between(60_000, 90_000);
         time::sleep(Duration::from_millis(millis)).await;
 
@@ -41,7 +41,7 @@ async fn execute(
     let groups_local = Group::find_all().await?;
 
     for failure in failures {
-        if failure.retry_count >= RauthyConfig::get().vars.scim.retry_count as i64 {
+        if failure.retry_count >= RauthyConfig::get().vars.scim.retry_count as i32 {
             warn!("Retry count exceeded for scim task {failure:?}");
 
             Event::scim_task_failed(&failure.client_id, &failure.action, failure.retry_count)
